+++ packages/furo-data/test/furo-custom-method.test.js
@@ -182,7 +182,7 @@ describe('furo-custom-method', () => {
     customMethod.trigger();
   });
 
-  it('Accept header should be set ', () => {
+  it('Accept header should be set ', done => {
     customMethod.setAttribute('service', 'ExperimentService');
     customMethod.setAttribute('method', 'createtemplate');
     dataObject.setAttribute('type', 'experiment.ExperimentEntity');
@@ -199,6 +199,7 @@ describe('furo-custom-method', () => {
         request.headers.get('Accept'),
         'application/experiment.Experiment+json, application/json;q=0.9',
       );
+      done();
     });
 
     customMethod.htsIn([
@@ -210,5 +211,135 @@ describe('furo-custom-method', () => {
         service: 'ExperimentService',
       },
     ]);
+  });
+
+  it('should accept QueryParams Object via updateQp method', done => {
+    customMethod.setAttribute('service', 'ExperimentService');
+    customMethod.setAttribute('method', 'createtemplate');
+
+    customMethod.updateQp({ compact: true });
+    customMethod.updateQp({ variant: '5' });
+
+    dataObject.setAttribute('type', 'experiment.ExperimentEntity');
+
+    /**
+     * Register hook on wire --triggerLoad to
+     *
+     */
+    customMethod._FBPAddWireHook('--triggerLoad', req => {
+      assert.equal(req.url.indexOf('compact=true') > 0, true);
+      assert.equal(req.url.indexOf('variant=5') > 0, true);
+      done();
+    });
+
+    customMethod.htsIn([
+      {
+        href: 'https://httpbin.org/anything',
+        method: 'Post',
+        rel: 'createtemplate',
+        type: 'experiment.ExperimentEntity',
+        service: 'ExperimentService',
+      },
+    ]);
+
+    customMethod.trigger();
+  });
+
+  it('should accept QueryParams Object via updateQp method', done => {
+    customMethod.setAttribute('service', 'ExperimentService');
+    customMethod.setAttribute('method', 'createtemplate');
+
+    customMethod.updateQp({ compact: true });
+    customMethod.updateQp({ variant: '5' });
+
+    dataObject.setAttribute('type', 'experiment.ExperimentEntity');
+
+    /**
+     * Register hook on wire --triggerLoad to
+     *
+     */
+    customMethod._FBPAddWireHook('--triggerLoad', req => {
+      assert.equal(req.url.indexOf('compact=true') > 0, true);
+      assert.equal(req.url.indexOf('variant=5') > 0, true);
+      done();
+    });
+
+    customMethod.htsIn([
+      {
+        href: 'https://httpbin.org/anything',
+        method: 'Post',
+        rel: 'createtemplate',
+        type: 'experiment.ExperimentEntity',
+        service: 'ExperimentService',
+      },
+    ]);
+
+    customMethod.trigger();
+  });
+
+  it('should accept clear params via clearQp method', done => {
+    customMethod.setAttribute('service', 'ExperimentService');
+    customMethod.setAttribute('method', 'createtemplate');
+
+    customMethod.updateQp({ compact: true });
+    customMethod.updateQp({ variant: '5' });
+    customMethod.clearQp();
+
+    dataObject.setAttribute('type', 'experiment.ExperimentEntity');
+
+    /**
+     * Register hook on wire --triggerLoad to
+     *
+     */
+    customMethod._FBPAddWireHook('--triggerLoad', req => {
+      assert.equal(req.url.indexOf('compact=true') > 0, false);
+      assert.equal(req.url.indexOf('variant=5') > 0, false);
+      done();
+    });
+
+    customMethod.htsIn([
+      {
+        href: 'https://httpbin.org/anything',
+        method: 'Post',
+        rel: 'createtemplate',
+        type: 'experiment.ExperimentEntity',
+        service: 'ExperimentService',
+      },
+    ]);
+
+    customMethod.trigger();
+  });
+
+  it('should rebuild request url with via updateQp setted qp and the previously existed qp together in url ', done => {
+    customMethod.setAttribute('service', 'ExperimentService');
+    customMethod.setAttribute('method', 'createtemplate');
+
+    customMethod.updateQp({ compact: true });
+    customMethod.updateQp({ variant: '5' });
+
+    dataObject.setAttribute('type', 'experiment.ExperimentEntity');
+
+    /**
+     * Register hook on wire --triggerLoad to
+     *
+     */
+    customMethod._FBPAddWireHook('--triggerLoad', req => {
+      assert.equal(req.url.indexOf('compact=true') > 0, true);
+      assert.equal(req.url.indexOf('variant=5') > 0, true);
+      assert.equal(req.url.indexOf('previousqp=xyz') > 0, true);
+      done();
+    });
+
+    customMethod.htsIn([
+      {
+        href: 'https://httpbin.org/anything?previousqp=xyz',
+        method: 'Post',
+        rel: 'createtemplate',
+        type: 'experiment.ExperimentEntity',
+        service: 'ExperimentService',
+      },
+    ]);
+
+    customMethod.trigger();
   });
 });