+++ packages/furo-ui5/demos/demos.js
@@ -1,0 +1,1 @@
+import './demo-furo-ui5-pagination-bar.js';