+++ packages/furo-ui5/src/furo-catalog.js
@@ -1,0 +1,1 @@
+import './furo-ui5-pagination-bar.js';