+++ packages/furo-icon/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.13](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/icon@1.3.12...@furo/icon@1.3.13) (2020-07-31)
+
+**Note:** Version bump only for package @furo/icon
+
+
+
+
+
 ## [1.3.12](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/icon@1.3.11...@furo/icon@1.3.12) (2020-06-29)
 
 **Note:** Version bump only for package @furo/icon