+++ packages/furo-input/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.16](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/input@1.3.15...@furo/input@1.3.16) (2020-07-31)
+
+**Note:** Version bump only for package @furo/input
+
+
+
+
+
 ## [1.3.15](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/input@1.3.14...@furo/input@1.3.15) (2020-06-29)
 
 **Note:** Version bump only for package @furo/input